% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_01_plan_qclist.R
\name{prepare_tasks_summary}
\alias{prepare_tasks_summary}
\title{Prepare the dataset}
\usage{
prepare_tasks_summary(df)
}
\arguments{
\item{df}{a dataframe with column names metodo, attivita, mese_previsto,
data_effettiva, operatore_previsto, operatore_effettivo, matrice, esito.}
}
\value{
a dataframe with the right class for each column.
}
\description{
prepare the dataset of for the tasks summary
}
